fix: remove AbortController test mock
[This change was recently made in rnmapbox.][0]

This library includes an `AbortController` test mock, which doesn't work
in places that expect a real one. For example, I was writing an
unrelated test that tried to cancel a `fetch()` which failed because the
`AbortController`'s `signal` property was invalid.

Because [`AbortController` is in all supported Node versions][1], we can
safely remove this mock and rely on the real thing.

This should only affect tests.

[0]: https://github.com/rnmapbox/maps/commit/5b6489e2bffe509ff26fbdc054b693682bcdb31e
[1]: https://developer.mozilla.org/en-US/docs/Web/API/AbortController#browser_compatibility